services/ml_service/requirements.txt
# GENERATOR: FULL_PLATFORM
# ML Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
faiss-cpu==1.7.4
gensim==4.3.2
python-dotenv==1.0.0
redis==5.0.1
requests>=2.31.0

